pyproject.toml
[tool.poetry]
name = "goit-web-hw-11"
version = "0.1.0"
description = ""
authors = ["Mizhik <vlad.mezhinskiy07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
redis = "^5.0.6"
cloudinary = "^1.40.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"